Rendering pipeline fixes across the board and base work on refactored editor
