+++ include/Engine/Editor/EditorUI.h
@@ -1,0 +1,8 @@
+#include <imgui/imgui.h>
+#include <nativefiledialog/nfd.h>
+
+class EditorUI
+{
+public:
+    EditorUI();
+};
