+++ include/Engine/Rendering/IRenderer.h
@@ -31,15 +31,6 @@ public:
 	void SetFullscreen(bool fullscreen) { m_Fullscreen = fullscreen; }
 	bool VSYNC() const { return m_VSYNC; }
 	void SetVSYNC(bool vsync) { m_VSYNC = vsync; }
-    ::Camera* Camera() const { return m_Camera; }
-    void SetCamera(::Camera* camera)
-    {
-        if (camera == nullptr) {
-            m_Camera = m_DefaultCamera;
-        } else {
-            m_Camera = camera;
-        }
-    }
 	virtual void Initialize() = 0;
 	virtual void Update(double dt) = 0;
 	virtual void Draw(RenderFrame& rq) = 0;
@@ -54,8 +45,6 @@ protected:
 	int m_GLVersion[2];
 	std::string m_GLVendor;
 	GLFWwindow* m_Window = nullptr;
-    ::Camera* m_DefaultCamera;
-    ::Camera* m_Camera = nullptr;
 };
 
 #endif // Renderer_h__
