+++ include/Engine/Rendering/RenderQueue.h
@@ -50,10 +50,11 @@ struct PointLightJob : RenderJob
 
 struct RenderScene
 {
-    ::Camera* Camera;
+    ::Camera* Camera = nullptr;
     std::list<std::shared_ptr<RenderJob>> ForwardJobs;
     std::list<std::shared_ptr<RenderJob>> PointLightJobs;
     Rectangle Viewport;
+    bool ClearDepth = false;
 
 	void Clear()
 	{
