+++ resources/Schema/Components/Camera.xsd
@@ -9,7 +9,6 @@
 		</xs:annotation>
 		<xs:complexType>
 			<xs:all>
-				<xs:element name="Name" type="t:string" minOccurs="0"/>
 				<xs:element name="FOV" type="t:double" minOccurs="0">
 					<xs:annotation><xs:documentation>Vertical Field of View in degrees</xs:documentation></xs:annotation>
 				</xs:element>
