+++ resources/Schema/Components/PointLight.xsd
@@ -12,7 +12,7 @@
 				<xs:element name="Color" type="t:Color" minOccurs="0"/>
 				<xs:element name="Radius" type="t:double" minOccurs="0"/>
 				<xs:element name="Intensity" type="t:double" minOccurs="0"/>
-				<xs:element name="Falloff">
+				<xs:element name="Falloff" minOccurs="0">
 					<xs:simpleType>
 						<xs:restriction base="t:double">
 							<xs:minInclusive value="0"/>
