+++ resources/Schema/Entities/EditorWidget.xml
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no" ?>
+<Entity xmlns:c="components" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../Types/Entity.xsd" name="TranslationWidget">
+
+	<Components>
+		<c:Transform/>
+		<c:UniformScale>
+			<Scale X="0.2" Y="0.2" Z="0.2"/>
+		</c:UniformScale>
+		
+	</Components>
+
+	<Children>
+		<Entity>
+			<Components>
+				<c:Transform>
+					<Position X="0.2" Y="0.2" Z="0.2"/>
+				</c:Transform>
+				<c:PointLight/>
+			</Components>
+		</Entity>
+		<Entity>
+			<Components>
+				<c:Transform/>
+				<c:Model>
+					<Resource>Models/TranslationWidgetOrigin.obj</Resource>
+				</c:Model>
+			</Components>
+		</Entity>
+		<Entity>
+			<Components>
+				<c:Transform/>
+				<c:Model>
+					<Resource>Models/TranslationWidgetX.obj</Resource>
+				</c:Model>
+			</Components>
+		</Entity>
+		<Entity>
+			<Components>
+				<c:Transform/>
+				<c:Model>
+					<Resource>Models/TranslationWidgetY.obj</Resource>
+				</c:Model>
+			</Components>
+		</Entity>
+		<Entity>
+			<Components>
+				<c:Transform/>
+				<c:Model>
+					<Resource>Models/TranslationWidgetZ.obj</Resource>
+				</c:Model>
+			</Components>
+		</Entity>
+		<Entity>
+			<Components>
+				<c:Transform/>
+				<c:Model>
+					<Resource>Models/WidgetPlaneX.obj</Resource>
+				</c:Model>
+			</Components>
+		</Entity>
+		<Entity>
+			<Components>
+				<c:Transform/>
+				<c:Model>
+					<Resource>Models/WidgetPlaneY.obj</Resource>
+				</c:Model>
+			</Components>
+		</Entity>
+		<Entity>
+			<Components>
+				<c:Transform/>
+				<c:Model>
+					<Resource>Models/WidgetPlaneZ.obj</Resource>
+				</c:Model>
+			</Components>
+		</Entity>
+	</Children>
+
+</Entity>
