+++ src/Engine/Rendering/RenderSystem.cpp
@@ -9,69 +9,24 @@ RenderSystem::RenderSystem(EventBroker* eventBroker, const IRenderer* renderer, 
     EVENT_SUBSCRIBE_MEMBER(m_EInputCommand, &RenderSystem::OnInputCommand);
 
     m_Camera = new Camera((float)m_Renderer->Resolution().Width / m_Renderer->Resolution().Height, glm::radians(45.f), 0.01f, 5000.f);
-    m_DebugCameraInputController = new DebugCameraInputController<RenderSystem>(eventBroker, -1);
 }
 
 RenderSystem::~RenderSystem()
 {
     delete m_Camera;
-    delete m_DebugCameraInputController;
 }
 
-bool RenderSystem::OnSetCamera(const Events::SetCamera &event)
+bool RenderSystem::OnSetCamera(Events::SetCamera& e)
 {
-    auto cameras = m_World->GetComponents("Camera");
-
-    if (cameras != nullptr) {
-        for (auto it = cameras->begin(); it != cameras->end(); it++) {
-            if ((std::string)(*it)["Name"] == event.Name) {
-                switchCamera((*it).EntityID);
-            }
-        }
-    }
+    ComponentWrapper cTransform = e.CameraEntity["Transform"];
+    ComponentWrapper cCamera = e.CameraEntity["Camera"];
+    m_Camera->SetFOV((double)cCamera["FOV"]);
+    m_Camera->SetNearClip((double)cCamera["NearClip"]);
+    m_Camera->SetFarClip((double)cCamera["FarClip"]);
+    m_Camera->SetPosition(cTransform["Position"]);
+    m_Camera->SetOrientation(glm::quat((const glm::vec3&)cTransform["Orientation"]));
+    m_CurrentCamera = e.CameraEntity;
     return true;
-}
-
-void RenderSystem::switchCamera(EntityID entity)
-{
-    if(m_World->HasComponent(entity, "Camera")) {
-
-        if (m_CurrentCamera != EntityID_Invalid) {
-            if (m_World->HasComponent(m_CurrentCamera, "Model")) {
-                m_World->GetComponent(m_CurrentCamera, "Model")["Visible"] = true;
-            }
-            if (m_World->HasComponent(m_CurrentCamera, "Listener")) {
-                m_World->DeleteComponent(m_CurrentCamera, "Listener");
-            }
-        }
-
-        if (m_World->HasComponent(entity, "Model")) {
-            m_World->GetComponent(entity, "Model")["Visible"] = false;
-        }
-        if (!m_World->HasComponent(entity, "Listener")) {
-            m_World->AttachComponent(entity, "Listener");
-        }
-        m_CurrentCamera = entity;
-        m_SwitchCamera = false;
-
-    } else {
-        LOG_ERROR("Entity %i does not have a CameraComponent", entity);
-        m_SwitchCamera = false;
-    }
-}
-
-void RenderSystem::updateProjectionMatrix(ComponentWrapper& cameraComponent)
-{
-    double fov = cameraComponent["FOV"];
-    double aspectRatio = (float)m_Renderer->Resolution().Width / m_Renderer->Resolution().Height;
-    double nearClip = cameraComponent["NearClip"];
-    double farClip = cameraComponent["FarClip"];
-
-    m_Camera->SetFOV(glm::radians(fov));
-    m_Camera->SetAspectRatio(aspectRatio);
-    m_Camera->SetNearClip(nearClip);
-    m_Camera->SetFarClip(farClip);
-    m_Camera->UpdateProjectionMatrix();
 }
 
 void RenderSystem::fillModels(std::list<std::shared_ptr<RenderJob>>& jobs, World* world)
@@ -113,7 +68,6 @@ void RenderSystem::fillModels(std::list<std::shared_ptr<RenderJob>>& jobs, World
     }
 }
 
-
 void RenderSystem::fillLight(std::list<std::shared_ptr<RenderJob>>& jobs, World* world)
 {
     auto pointLights = world->GetComponents("PointLight");
@@ -138,12 +92,7 @@ void RenderSystem::fillLight(std::list<std::shared_ptr<RenderJob>>& jobs, World*
 
 bool RenderSystem::OnInputCommand(const Events::InputCommand& e)
 {
-    if (e.Command == "SwitchCamera" && e.Value > 0) {
-        m_SwitchCamera = true;
-        return true;
-    } else {
-        return false;
-    }
+    return false;
 }
 
 void RenderSystem::Update(World* world, double dt)
@@ -151,10 +100,12 @@ void RenderSystem::Update(World* world, double dt)
     m_World = world;
     m_EventBroker->Process<RenderSystem>();
 
-    updateCamera(world, dt);
-
+    if (m_CurrentCamera) {
+        ComponentWrapper cameraTransform = m_CurrentCamera["Transform"];
+        m_Camera->SetPosition(cameraTransform["Position"]);
+        m_Camera->SetOrientation(glm::quat((const glm::vec3&)cameraTransform["Orientation"]));
+    }
     //Only supports opaque geometry atm
-    m_RenderFrame->Clear();
 
     RenderScene scene;
     scene.Camera = m_Camera;
@@ -163,70 +114,4 @@ void RenderSystem::Update(World* world, double dt)
     fillLight(scene.PointLightJobs, world);
     m_RenderFrame->Add(scene);
    
-}
-
-void RenderSystem::updateCamera(World* world, double dt)
-{
-    if (m_SwitchCamera) {
-        auto cameras = world->GetComponents("Camera");
-        if (cameras == nullptr) {
-            return;
-        }
-        for (auto it = cameras->begin(); it != cameras->end(); it++) {
-            if ((*it).EntityID == m_CurrentCamera) {
-                it++;
-                if (it != cameras->end()) {
-                    switchCamera((*it).EntityID);
-                } else {
-                    switchCamera((*cameras->begin()).EntityID);
-                }
-                break;
-            }
-        }
-        if (m_World->HasComponent(m_CurrentCamera, "Camera")) {
-            ComponentWrapper& cameraComponent = world->GetComponent(m_CurrentCamera, "Camera");
-            ComponentWrapper& cameraTransform = world->GetComponent(m_CurrentCamera, "Transform");
-
-            m_DebugCameraInputController->SetOrientation(glm::quat((glm::vec3)cameraTransform["Orientation"]));
-            m_DebugCameraInputController->SetPosition(cameraTransform["Position"]);
-        }
-    }
-
-    if (m_World->ValidEntity(m_CurrentCamera)) {
-        if (world->HasComponent(m_CurrentCamera, "Camera") && world->HasComponent(m_CurrentCamera, "Transform")) {
-            ComponentWrapper& cameraComponent = world->GetComponent(m_CurrentCamera, "Camera");
-            ComponentWrapper& cameraTransform = world->GetComponent(m_CurrentCamera, "Transform");
-
-            m_DebugCameraInputController->Update(dt);
-            (glm::vec3&)cameraTransform["Orientation"] = glm::eulerAngles(m_DebugCameraInputController->Orientation());
-            (glm::vec3&)cameraTransform["Position"] = m_DebugCameraInputController->Position();
-
-            glm::vec3 position = Transform::AbsolutePosition(world, m_CurrentCamera);
-            glm::quat orientation = Transform::AbsoluteOrientation(world, m_CurrentCamera);
-
-            m_Camera->SetPosition(position);
-            m_Camera->SetOrientation(orientation);
-
-            updateProjectionMatrix(cameraComponent);
-
-        }
-    } else {
-        m_Camera = m_Camera;
-
-        auto cameras = world->GetComponents("Camera");
-        if (cameras != nullptr) {
-            if (cameras->begin() != cameras->end()) {
-                ComponentWrapper& cameraC = *cameras->begin();
-                switchCamera(cameraC.EntityID);
-
-                ComponentWrapper& cameraComponent = world->GetComponent(m_CurrentCamera, "Camera");
-                ComponentWrapper& cameraTransform = world->GetComponent(m_CurrentCamera, "Transform");
-
-                m_DebugCameraInputController->SetOrientation(glm::quat((glm::vec3)cameraTransform["Orientation"]));
-                m_DebugCameraInputController->SetPosition(cameraTransform["Position"]);
-            }
-        }
-    }
-
-    m_Camera->UpdateViewMatrix();
 }
